+++ src/test/converter2/behavior/categoryInheritance.ts
@@ -1,0 +1,8 @@
+export interface Int {
+    /** @category Cat */
+    prop: string;
+}
+
+export class Cls implements Int {
+    prop = "";
+}