+++ src/test/converter2/behavior/groupInheritance.ts
@@ -1,0 +1,8 @@
+export interface Int {
+    /** @group Group */
+    prop: string;
+}
+
+export class Cls implements Int {
+    prop = "";
+}